refactor: all safe globals in one place
