+++ stable_baselines3/common/off_policy_algorithm.py
@@ -228,7 +228,7 @@ class OffPolicyAlgorithm(BaseAlgorithm):
         self,
         path: str | pathlib.Path | io.BufferedIOBase,
         truncate_last_traj: bool = True,
-        deserialization_mode: str = "legacy",
+        deserialization_mode: str = "safe",
     ) -> None:
         """
         Load a replay buffer from a pickle file.
@@ -240,10 +240,13 @@ class OffPolicyAlgorithm(BaseAlgorithm):
             If set to ``False``, we assume that we continue the same trajectory (same episode).
         :param deserialization_mode: How to handle pickle deserialization.
 
-            - ``"legacy"`` (default): Deserialize with ``pickle.load()``.  This
-              preserves backward compatibility but can **execute arbitrary
-              Python code** embedded in the pickle file.
-            - ``"safe"``: Refuse to deserialize.  Raises a ``ValueError``.
+            - ``"safe"`` (default): Deserialize using a restricted unpickler that
+              only allows a fixed allowlist of known-safe SB3/gymnasium/numpy types.
+              Any pickle payload referencing a type outside this allowlist is
+              rejected with a clear error.
+            - ``"legacy"``: Deserialize with ``pickle.load()``.  This preserves
+              backward compatibility but can **execute arbitrary Python code**
+              embedded in the pickle file.
 
         """
         self.replay_buffer = load_from_pkl(
